+++ src/_frame_test.js
@@ -222,6 +222,17 @@ describe("Frame", function() {
 			assert.equal(frameDom.contentDocument.body.innerHTML, "", "frame body");
 		});
 
+		it.only("scrolls", function() {
+			// put an element outside the window bounds so the window can be scrolled
+			frame.addElement("<div style='position: absolute; left: 5000px; top: 5000px; width: 60px'>element</div>");
+
+			assert.deepEqual(frame.getRawScrollPosition(), { x: 0, y: 0}, "should start at (0, 0)");
+
+			frame.scroll(150, 300);
+			assert.equal(frame.getRawScrollPosition().x, 150, "should have scrolled right");
+			assert.equal(frame.getRawScrollPosition().y, 300, "should have scrolled down");
+		});
+
 	});
 
 });